requirements.txt
numpy
pandas
sklearn
flask
flask-cors
gunicorn
PyYAML
dill
-e .